Sign the `tree` executable.
Added a code signing step to the build.  The stamp file
`$(TMP)/signed.stamp.txt` tracks the signing state of the `tree`
executable after it is installed for packaging.

The makefile variable `APP_SIGNING_ID` holds the name of the Developer
ID Application cert used to sign `tree`.

Also added some notes on this variable, the `TMP` variable and build
commands to the readme file.