Add a `notarize` target to the makefile.
Rename the `all` target to `signed-package`, since it really doesn't do
"all".  Add new target `notarize` with steps to notarize the signed
installer package and staple the notarization ticket to it.  The build
creates intermediate files:
- `$(TMP)/submit-log.json` tracks notarytool submission
- `$(TMP)/submission-id.txt` holds the submission ID extracted from the
  submit log
- `$(TMP)/notarization-log.json` holds the full notarization report
- `$(TMP)/notarized.stamp.txt` tracks "Accepted" (success) status
- `$(TMP)/stapled.stamp.txt` tracks stapling status

Also added makefile variable `NOTARIZATION_KEYCHAIN_PROFILE`, with a
default of "Donald McCaughey", for the name of the Keychain item holding
notarization service credentials.

Also updated the readme file with a section on prerequisites, info on
the `NOTARIZATION_KEYCHAIN_PROFILE` variable and a section on sigining
and notarizing credentials.