Update `tree` from 1.8.0 to 2.0.1.
Pull in the source for `tree` 2.0.1.  Update version numbers in
`Makefile` and `README.md`.  Update the macOS build environment
description to match the latest macOS and Xcode.

Modified dist/Makefile to uncomment macOS settings as per the
instructions in the makefile, and to modify the use of the `DISTDIR`
variable in the install target so that installation for packaging works
as expected.