Fix broken dependency tree for the `notarize` target.
Since the `stapler` command modifies the notarized package in place,
out-of-date resolution for make is broken.  Fix this by renaming the
unnotarized package from `tree-$(ver).pkg` to
`$(TMP)/tree-$(ver)-unnotarized.pkg` and changing the stapling step to
first copy the package to `tree-$(ver).pkg` then stapling it.